refact: Rename statistics app files
