+++ projects/test-statistics-app/src/app/app.component.html
